README.md
The Ferris Wheel project is a fun and interactive web application to provide users with a virtual Ferris Wheel experience. It was developed as a learning project to explore HTML, CSS, and animations, to create a visually appealing and engaging user interface.

This project suits anyone interested in web development, animations, and creative design. Users can enjoy watching the rotating wheel and cabins and experiencing the smooth animation effects and color transitions implemented using CSS.